move swarm-scenario:Swarm.Game.Failure to swarm-util:Swarm.Failure
