+++ app/game/Swarm/App.hs
@@ -32,7 +32,7 @@ import Data.Text.IO qualified as T
 import GitHash (GitInfo)
 import Graphics.Vty qualified as V
 import Graphics.Vty.CrossPlatform qualified as V
-import Swarm.Game.Failure (SystemFailure)
+import Swarm.Failure (SystemFailure)
 import Swarm.Game.State.Runtime
 import Swarm.Log (LogSource (SystemLog), Severity (..))
 import Swarm.Pretty (prettyText)
