+++ src/swarm-engine/Swarm/Game/Achievement/Persistence.hs
@@ -15,7 +15,7 @@ import Data.Sequence (Seq)
 import Data.Yaml qualified as Y
 import Swarm.Game.Achievement.Attainment
 import Swarm.Game.Achievement.Definitions
-import Swarm.Game.Failure
+import Swarm.Failure
 import Swarm.Game.ResourceLoading (getSwarmAchievementsPath)
 import Swarm.Util.Effect (forMW)
 import System.Directory (doesDirectoryExist, doesFileExist, listDirectory)
