+++ src/swarm-scenario/Swarm/Game/Land.hs
@@ -17,7 +17,7 @@ import Control.Effect.Throw (Throw)
 import Control.Lens (makeLenses)
 import GHC.Generics (Generic)
 import Swarm.Game.Entity
-import Swarm.Game.Failure (SystemFailure)
+import Swarm.Failure (SystemFailure)
 import Swarm.Game.Terrain
 
 data TerrainEntityMaps = TerrainEntityMaps
