+++ src/swarm-scenario/Swarm/Game/ResourceLoading.hs
@@ -35,7 +35,7 @@ import Data.Maybe (mapMaybe)
 import Data.Text (Text)
 import Data.Text qualified as T
 import Paths_swarm (getDataDir)
-import Swarm.Game.Failure
+import Swarm.Failure
 import Swarm.Util
 import System.Directory (
   XdgDirectory (..),
