+++ src/swarm-scenario/Swarm/Game/Terrain.hs
@@ -37,7 +37,7 @@ import Data.Yaml
 import GHC.Generics (Generic)
 import Swarm.Game.Display
 import Swarm.Game.Entity.Cosmetic (WorldAttr (..))
-import Swarm.Game.Failure
+import Swarm.Failure
 import Swarm.Game.ResourceLoading (getDataFileNameSafe)
 import Swarm.Util (enumeratedMap, quote)
 import Swarm.Util.Effect (withThrow)
