+++ src/swarm-tui/Swarm/TUI/Launch/Model.hs
@@ -15,7 +15,7 @@ import Control.Carrier.Throw.Either (runThrow)
 import Control.Lens (makeLenses)
 import Data.Functor.Identity (Identity (Identity))
 import Data.Text (Text)
-import Swarm.Game.Failure (SystemFailure)
+import Swarm.Failure (SystemFailure)
 import Swarm.Game.Scenario.Status (ParameterizableLaunchParams (LaunchParams), ScenarioInfoPair, SerializableLaunchParams)
 import Swarm.Game.State (LaunchParams, ValidatedLaunchParams, getRunCodePath, parseCodeFile)
 import Swarm.Pretty (prettyText)
