+++ src/swarm-tui/Swarm/TUI/Model/KeyBindings.hs
@@ -22,7 +22,7 @@ import Data.Bifunctor (second)
 import Data.Maybe (fromMaybe, mapMaybe)
 import Data.Text (Text)
 import Data.Text qualified as T
-import Swarm.Game.Failure (Asset (..), LoadingFailure (..), SystemFailure (..))
+import Swarm.Failure (Asset (..), LoadingFailure (..), SystemFailure (..))
 import Swarm.Game.ResourceLoading (getSwarmConfigIniFile)
 import Swarm.Pretty (prettyText)
 import Swarm.TUI.Controller.EventHandlers
